Automatically backed up by Learn

+++ lib/cash_register.rb
@@ -17,7 +17,7 @@ class CashRegister
     
   def apply_discount()
     if @discount > 0 
-      @discount = @discount/100.to_f 
+      @discount = @discount/100.to_i 
       @total = @total - (@total * @discount)
       "After the discount, the total comes to $#{@total}."
     end
